+++ packages/dart/noports_core/test/srvd/notification_subscription_test.dart
@@ -3,6 +3,8 @@ import 'dart:io';
 
 import 'package:at_client/at_client.dart';
 import 'package:mocktail/mocktail.dart';
+import 'package:noports_core/src/srvd/relay_auth_verifiers.dart'
+    show defaultRelayAuthDetectWindowMs;
 import 'package:noports_core/src/srvd/srvd_impl.dart';
 import 'package:noports_core/srvd.dart';
 import 'package:test/test.dart';
@@ -71,6 +73,7 @@ void main() {
         verbose: false,
         bind443: false,
         localBindPort443: 443,
+        relayAuthDetectWindowMs: defaultRelayAuthDetectWindowMs,
       );
 
       // Create a stream controller to simulate the notification received from the sshnp